cmd/testutils/combinationtest: Create visual representation test utility.
Since combination algorithms use random which makes it hard to write unit tests
for them, so it is easier to write a visualizing application that shows what
happens on small test cases. Later we can upgrade to unit tests when we create
our own random source which can have test values.